Move generator_pb2 and sequence_generator out of music
