+++ magenta/interfaces/midi/midi_interaction.py
@@ -19,7 +19,7 @@ import threading
 import time
 
 import magenta
-from magenta.music.protobuf import generator_pb2
+from magenta.protobuf import generator_pb2
 from magenta.music.protobuf import music_pb2
 import tensorflow as tf
