+++ magenta/models/improv_rnn/improv_rnn_model.py
@@ -147,7 +147,7 @@ class ImprovRnnConfig(events_rnn_model.EventSequenceRnnConfig):
 default_configs = {
     'basic_improv':
         ImprovRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='basic_improv',
                 description='Basic melody-given-chords RNN with one-hot triad '
                 'encoding for chords.'),
@@ -165,7 +165,7 @@ default_configs = {
                 learning_rate=0.001)),
     'attention_improv':
         ImprovRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='attention_improv',
                 description='Melody-given-chords RNN with one-hot triad encoding '
                 'for chords, attention, and binary counters.'),
@@ -183,7 +183,7 @@ default_configs = {
                 learning_rate=0.001)),
     'chord_pitches_improv':
         ImprovRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='chord_pitches_improv',
                 description='Melody-given-chords RNN with chord pitches encoding.'
             ),
