+++ magenta/models/performance_rnn/performance_model.py
@@ -187,7 +187,7 @@ class PerformanceRnnConfig(events_rnn_model.EventSequenceRnnConfig):
 default_configs = {
     'performance':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='performance', description='Performance RNN'),
             magenta.music.OneHotEventSequenceEncoderDecoder(
                 magenta.music.PerformanceOneHotEncoding()),
@@ -199,7 +199,7 @@ default_configs = {
                 learning_rate=0.001)),
     'performance_with_dynamics':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='performance_with_dynamics',
                 description='Performance RNN with dynamics'),
             magenta.music.OneHotEventSequenceEncoderDecoder(
@@ -213,7 +213,7 @@ default_configs = {
             num_velocity_bins=32),
     'performance_with_dynamics_compact':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='performance_with_dynamics',
                 description='Performance RNN with dynamics (compact input)'),
             magenta.music.OneHotIndexEventSequenceEncoderDecoder(
@@ -227,7 +227,7 @@ default_configs = {
             num_velocity_bins=32),
     'performance_with_dynamics_and_modulo_encoding':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='performance_with_dynamics_and_modulo_encoding',
                 description='Performance RNN with dynamics and modulo encoding'
             ),
@@ -242,7 +242,7 @@ default_configs = {
             num_velocity_bins=32),
     'performance_with_dynamics_and_note_encoding':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='performance_with_dynamics_and_note_encoding',
                 description='Performance RNN with dynamics and note encoding'),
             magenta.music.NotePerformanceEventSequenceEncoderDecoder(
@@ -257,7 +257,7 @@ default_configs = {
             note_performance=True),
     'density_conditioned_performance_with_dynamics':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='density_conditioned_performance_with_dynamics',
                 description='Note-density-conditioned Performance RNN + dynamics'
             ),
@@ -277,7 +277,7 @@ default_configs = {
             ]),
     'pitch_conditioned_performance_with_dynamics':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='pitch_conditioned_performance_with_dynamics',
                 description='Pitch-histogram-conditioned Performance RNN'),
             magenta.music.OneHotEventSequenceEncoderDecoder(
@@ -295,7 +295,7 @@ default_configs = {
             ]),
     'multiconditioned_performance_with_dynamics':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='multiconditioned_performance_with_dynamics',
                 description='Density- and pitch-conditioned Performance RNN'),
             magenta.music.OneHotEventSequenceEncoderDecoder(
@@ -316,7 +316,7 @@ default_configs = {
             ]),
     'optional_multiconditioned_performance_with_dynamics':
         PerformanceRnnConfig(
-            magenta.music.protobuf.generator_pb2.GeneratorDetails(
+            magenta.protobuf.generator_pb2.GeneratorDetails(
                 id='optional_multiconditioned_performance_with_dynamics',
                 description='Optionally multiconditioned Performance RNN'),
             magenta.music.OneHotEventSequenceEncoderDecoder(
