+++ magenta/models/performance_rnn/performance_rnn_generate.py
@@ -25,7 +25,7 @@ import magenta
 from magenta.models.performance_rnn import performance_model
 from magenta.models.performance_rnn import performance_sequence_generator
 from magenta.music import constants
-from magenta.music.protobuf import generator_pb2
+from magenta.protobuf import generator_pb2
 from magenta.music.protobuf import music_pb2
 import tensorflow as tf
