+++ magenta/models/pianoroll_rnn_nade/README.md
@@ -153,7 +153,7 @@ is a convenient way of combining the model checkpoint, metagraph, and
 some metadata about the model into a single file.
 
 To generate a bundle, use the
-[create_bundle_file](/magenta/music/sequence_generator.py)
+[create_bundle_file](/magenta/models/shared/sequence_generator.py)
 method within SequenceGenerator. Our generator script
 supports a ```--save_generator_bundle``` flag that calls this method. Example:
