+++ magenta/models/pianoroll_rnn_nade/pianoroll_rnn_nade_model.py
@@ -17,7 +17,7 @@
 from magenta.models.pianoroll_rnn_nade import pianoroll_rnn_nade_graph
 from magenta.models.shared import events_rnn_model
 import magenta.music as mm
-from magenta.music.protobuf import generator_pb2
+from magenta.protobuf import generator_pb2
 import tensorflow as tf
