+++ magenta/models/polyphony_rnn/README.md
@@ -161,7 +161,7 @@ The [bundle format](/magenta/protobuf/generator.proto)
 is a convenient way of combining the model checkpoint, metagraph, and
 some metadata about the model into a single file.
 
-To generate a bundle, use the [create_bundle_file](/magenta/music/sequence_generator.py) method within SequenceGenerator. Our generator script supports a `--save_generator_bundle` flag that calls this method. When using the `--save_generator_bundle` mode, you need to supply the `--hparams` flag with the same values used during training.
+To generate a bundle, use the [create_bundle_file](/magenta/models/shared/sequence_generator.py) method within SequenceGenerator. Our generator script supports a `--save_generator_bundle` flag that calls this method. When using the `--save_generator_bundle` mode, you need to supply the `--hparams` flag with the same values used during training.
 
 Example:
