+++ magenta/models/polyphony_rnn/polyphony_model.py
@@ -17,7 +17,7 @@
 import magenta
 from magenta.models.polyphony_rnn import polyphony_encoder_decoder
 from magenta.models.shared import events_rnn_model
-from magenta.music.protobuf import generator_pb2
+from magenta.protobuf import generator_pb2
 import tensorflow as tf
