+++ magenta/protobuf/generator.proto
